refactor: 이벤트 등록 부분 구조 변경
