+++ src/js/App.js
@@ -1,63 +1,14 @@
-import { $ } from './utils/querySelector.js';
-import throttle, { SCROLL_THROTTLE_DELAY } from './utils/throttle.js';
-import debounce from './utils/debounce.js';
-import { handleSearch, handleScrollSearch } from './handlers/searchHandle.js';
-import { handleSaveButtonClick } from './handlers/saveVideoHandle.js';
-import {
-  handleGonnaWatchToggleClick,
-  handleWatchedToggleClick,
-} from './handlers/showSavedVideosHandle.js';
-import { handleModalClose } from './handlers/modalHandle.js';
-import {
-  handleWatchedButtonClick,
-  handleDeleteButtonClick,
-} from './handlers/savedVideoButtonHandle.js';
-import mainPageUI from './ui/mainPage/mainPageUI.js';
+import mainPageUI from './views/mainPage/mainPageUI.js';
 import videoStorage from './localStorage/videoStorage.js';
+import bindMainPageEvents from './views/mainPage/mainPageEvents.js';
+import bindModalEvents from './views/modal/modalEvents.js';
 
 export default function App() {
+  bindMainPageEvents();
+  bindModalEvents();
+
   const savedVideos = videoStorage.getSavedVideos();
   mainPageUI.renderSavedVideoItems(savedVideos);
-
-  // 검색 이벤트
-  $('.search-input').addEventListener('submit', e => {
-    e.preventDefault();
-    $('.suggestion').hidden = true;
-    handleSearch();
-  });
-
-  // 스크롤 이벤트
-  $('.video-list').addEventListener('scroll', throttle(handleScrollSearch, SCROLL_THROTTLE_DELAY));
-
-  // 저장 버튼 클릭 이벤트
-  $('.video-list').addEventListener('click', handleSaveButtonClick);
-
-  // 검색 버튼 클릭 이벤트
-  $('#search-modal-button').addEventListener('click', () => {
-    $('.modal-container').classList.toggle('hide');
-  });
-
-  // 모달 외의 영역 클릭 이벤트
-  $('.dimmer').addEventListener('click', handleModalClose);
-
-  // 볼 영상 버튼 클릭
-  $('#gonna-watch-button').addEventListener('click', handleGonnaWatchToggleClick);
-
-  // 본 영상 버튼 클릭
-  $('#watched-button').addEventListener('click', handleWatchedToggleClick);
-
-  // 영상 저장, 삭제 버튼
-  $('.saved-video-list').addEventListener('click', e => {
-    const isWatchedButtonClick = e.target.classList.contains('video-item__watched-button');
-    const isDeleteButtonClick = e.target.classList.contains('video-item__delete-button');
-
-    if (isWatchedButtonClick) {
-      handleWatchedButtonClick(e);
-    }
-    if (isDeleteButtonClick) {
-      handleDeleteButtonClick(e);
-    }
-  });
 }
 ///////////////////// 검색어 추천
