+++ src/js/handlers/showSavedVideosHandle.js
@@ -1,5 +1,5 @@
 import { $ } from '../utils/querySelector.js';
-import mainPageUI from '../ui/mainPage/mainPageUI.js';
+import mainPageUI from '../views/mainPage/mainPageUI.js';
 import videoStorage from '../localStorage/videoStorage.js';
 
 export const handleGonnaWatchToggleClick = () => {
