+++ src/js/views/mainPage/mainPageUI.js
