+++ src/js/views/mainPage/template.js
