+++ src/js/views/modal/template.js
